requirements.txt
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openpyxl==3.1.2
pandas==2.3.3
python-dotenv==1.0.1
pytest==8.3.3
pytest-mock==3.14.0
